pthreads libxtrans fontconfig pixman xserver mesa plink xkeyboard-config libX11 git update Nov 6 2019
xserver          commit a506b4ecb6c54fd0388e628520eec75ea3bcb27c
xkeyboard-config commit 08fffa61190fcf8478e89b0d95e326bac527b7f3
libX11           commit b8766a438461e20b5c6f7a6111b703d4cb71501f
pixman           commit 85acb0a933bffbce6af03339546aff0a1eab5096
libxtrans        commit 3b5df889f58a99980a35a7b4a18eb4e7d2abeac4
fontconfig       commit 23ab038281fe59476a85bee02ad0ceb419095bdb
mesa             commit c1c346f1667375e9330aa3f729b97e4a6ece0320
putty            commit 2e93c3868e012c7a7750dea89874b70c96632fde
pthreads         commit 67186579ee3cf81e54d149dce5f5a612f957d2fe

+++ mesalib/src/gallium/auxiliary/util/u_format_zs.c
@@ -45,7 +45,7 @@ static inline uint32_t
 z16_unorm_to_z32_unorm(uint16_t z)
 {
    /* z * 0xffffffff / 0xffff */
-   return (z << 16) | z;
+   return ((uint32_t)z << 16) | z;
 }
 
 static inline uint32_t
@@ -439,7 +439,7 @@ util_format_z24_unorm_s8_uint_pack_s_8uint(uint8_t *dst_row, unsigned dst_stride
       for(x = 0; x < width; ++x) {
          uint32_t value = util_le32_to_cpu(*dst);
          value &= 0x00ffffff;
-         value |= *src++ << 24;
+         value |= (uint32_t)*src++ << 24;
          *dst++ = util_cpu_to_le32(value);
       }
       dst_row += dst_stride/sizeof(*dst_row);
@@ -459,7 +459,7 @@ util_format_z24_unorm_s8_uint_pack_separate(uint8_t *dst_row, unsigned dst_strid
       const uint8_t *s_src = s_src_row;
       uint32_t *dst = (uint32_t *)dst_row;
       for (x = 0; x < width; ++x) {
-         *dst++ = (*z_src++ & 0x00ffffff) | (*s_src++ << 24);
+         *dst++ = (*z_src++ & 0x00ffffff) | ((uint32_t)*s_src++ << 24);
       }
       dst_row += dst_stride / sizeof(*dst_row);
       z_src_row += z_src_stride / sizeof(*z_src_row);
